fix book added update or delete req

+++ server/app/controllers/book/DeleteBook.ts
@@ -4,9 +4,18 @@ import httpStatus from 'http-status'
 import sendResponse from '../../../utils/SendResponse';
 import { IBook } from '../../../interfaces/modelTypes';
 import Book from '../../models/BookModel';
+import ApiError from '../../../utils/errors/ApiError';
 
 const DeleteBook: RequestHandler = catchAsync(
     async (req: Request, res: Response) => {
+
+        // finding the book
+        const findBook = await Book.findById(req.params.id);
+
+        // checking user req is valid
+        if (findBook?.addedBy.toString() !== req?.user?._id) {
+            throw new ApiError(httpStatus.BAD_REQUEST, 'You have not added this book!')
+        }
 
         // deleting specific cow
         const result = await Book.findByIdAndDelete(req.params.id);
